+++ chatbot.js
@@ -430,7 +430,7 @@ function onRoomState(channel, state) {
     countLoadedChannels++;
     if (!connected && countLoadedChannels === config.channels.length) {
         connected = true;
-        console.log(locales.t('channels-loaded'));
+        console.log(locales.t('channels-loaded', [config.channels.join(', ')]));
 
         // wait for chatbot.warmUpDatabase(state)
         setTimeout(function() {