+++ src/js/vue/app/fontawesome.js
@@ -26,6 +26,7 @@ import {
     faDragon as fasDragon,
     faEdit as fasEdit,
     faExternalLinkAlt as fasExternalLinkAlt,
+    faEye as fasEye,
     faGavel as fasGavel,
     faGem as fasGem,
     faGift as fasGift,
@@ -59,7 +60,7 @@ library.add(fabEthereum, fabFontAwesomeFlag, fabTwitch, fabYoutube);
 library.add(farSadCry, farQuestionCircle);
 library.add(
     fasAward, fasBolt, fasBug, fasChartPie, fasCheckCircle, fasChevronRight, fasCogs, 
-    fasCommentDots, fasCopy, fasCrown, fasDragon, fasEdit, fasExternalLinkAlt, fasGavel, 
+    fasCommentDots, fasCopy, fasCrown, fasDragon, fasEdit, fasExternalLinkAlt, fasEye, fasGavel, 
     fasGem, fasGift, fasGlobe, fasHammer, fasHdd, fasHome, fasKey, fasMoneyBill, 
     fasPlay, fasRobot, fasSave, fasSnowflake, fasStar, fasStepForward, fasSync, 
     fasTerminal, fasTh, fasTicketAlt, fasTimes, fasTrain, fasTrashAlt,